local dir support
