+++ index.html
@@ -52,6 +52,11 @@
                 </div>
                 <input type="text" id="accessToken" name="accessToken" class="mt-1 block w-full rounded-md border border-gray-300 shadow-sm focus:border-indigo-300 focus:ring focus:ring-indigo-200 focus:ring-opacity-50 h-10 px-2">
             </div>
+            <div>
+                <a href="/local.html" class="text-blue-500 hover:text-blue-700 underline">
+                    Convert From Local Directory Instead
+                </a>
+            </div>
             <button type="submit" class="bg-blue-500 hover:bg-blue-600 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline flex items-center justify-center">
                 <i data-lucide="folder-search" class="w-5 h-5 mr-2"></i>
                 Fetch Directory Structure
