+++ local.html
@@ -54,9 +54,10 @@
             <input type="file" id="directoryPicker" webkitdirectory directory multiple class="mt-1 block w-full">
         </div>
 
-        <div class="mb-4 hidden">
-            <label for="gitignoreContent" class="block text-sm font-medium text-gray-600">.gitignore:</label>
-            <textarea id="gitignoreContent" rows="5" class="mt-1 block w-full rounded-md border border-gray-300 shadow-sm focus:border-indigo-300 focus:ring focus:ring-indigo-200 focus:ring-opacity-50 p-2"></textarea>
+        <div>
+            <a href="/" class="text-blue-500 hover:text-blue-700 underline">
+                Convert From GitHub Instead
+            </a>
         </div>
 
         <div id="extentionCheckboxes" class="mt-4"></div>
